pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "goforge"
version = "0.1.1"
description = "generate entire CRUD backend for golang echo and postgres with one command"
authors = ["Prashant Bhatkal <prashantbhatkal2000@gmail.com>"]
readme = "README.md"
keywords = [
    "code generator",
	"golang echo",
	"postgres"
]
packages = [{ include = "goforge", from = "." }]

[tool.poetry.dependencies]
python = "^3.7"
Jinja2 = ">=3.1.2"
pyyaml = ">=5.4.1"

[tool.poetry.scripts]
goforge = "goforge.main:entrypoint"

